+++ app/models/user.rb
@@ -40,6 +40,7 @@ class User < ActiveRecord::Base
   has_many :datasets, -> { where deleted: false }
   has_many :dataset_file_audits
   has_many :hosting_requests, -> { current }
+  has_many :images
   has_many :reviews
   has_many :tags, -> { where deleted: false }
   has_many :tools